+++ src/Stitching/Stitching.Tests/Middleware/DelegateToRemoteSchemaMiddlewareTests.cs
@@ -1,3 +1,5 @@
+using System.Text;
+using System.IO;
 using System.Linq;
 using System.Security.Cryptography.X509Certificates;
 using System.Net.Cache;
@@ -527,7 +529,7 @@ namespace HotChocolate.Stitching
             Snapshot.Match(result);
         }
 
-        [Fact(Skip = "Fix this issue")]
+        [Fact]
         public async Task ExtendedScalarAsInAndOutputType()
         {
             // arrange
@@ -611,7 +613,57 @@ namespace HotChocolate.Stitching
                 }");
                 request.VariableValues = new Dictionary<string, object>
                 {
-                    {"d", "2019-01-01T01:00"}
+                    {"d", "2019-01-01T01:00:00.000Z"}
+                };
+                request.Services = scope.ServiceProvider;
+
+                result = await executor.ExecuteAsync(request);
+            }
+
+            // assert
+            Snapshot.Match(result);
+        }
+
+        [Fact]
+        public async Task DateTimeIsHandledCorrectly()
+        {
+            // arrange
+            IHttpClientFactory clientFactory = CreateRemoteSchemas();
+
+            var serviceCollection = new ServiceCollection();
+            serviceCollection.AddSingleton(clientFactory);
+            serviceCollection.AddStitchedSchema(builder =>
+                builder.AddSchemaFromHttp("contract")
+                    .AddSchemaFromHttp("customer")
+                    .AddExtensionsFromString(
+                        "directive @custom(d: DateTime) on FIELD")
+                    .AddSchemaConfiguration(c =>
+                    {
+                        c.RegisterExtendedScalarTypes();
+                    })
+                    .AddSchemaConfiguration(c =>
+                        c.RegisterType<PaginationAmountType>()));
+
+            IServiceProvider services =
+                serviceCollection.BuildServiceProvider();
+
+            IQueryExecutor executor = services
+                .GetRequiredService<IQueryExecutor>();
+            IExecutionResult result = null;
+
+            // act
+            using (IServiceScope scope = services.CreateScope())
+            {
+                var request = new QueryRequest(@"
+                query a($d: DateTime!) {
+                    a: extendedScalar(d: ""2018-01-01T01:00:00.000Z"")
+                    b: extendedScalar(d: $d)
+                    c: extendedScalar(d: $d)
+                        @custom(d: ""2020-09-01T01:00:00.000Z"")
+                }");
+                request.VariableValues = new Dictionary<string, object>
+                {
+                    {"d", "2019-01-01T01:00:00.000Z"}
                 };
                 request.Services = scope.ServiceProvider;
 
@@ -886,6 +938,57 @@ namespace HotChocolate.Stitching
             IExecutionResult result =
                 await ExecutedMutationWithRenamedInputField(
                     requestBuilder);
+
+            // assert
+            Snapshot.Match(result);
+        }
+
+        [Fact]
+        public async Task AddErrorFilter()
+        {
+            // arrange
+            IHttpClientFactory clientFactory = CreateRemoteSchemas();
+
+            var serviceCollection = new ServiceCollection();
+            serviceCollection.AddSingleton(clientFactory);
+            serviceCollection.AddStitchedSchema(builder =>
+                builder.AddSchemaFromHttp("contract")
+                    .AddSchemaFromHttp("customer")
+                    .AddExtensionsFromString(
+                        FileResource.Open("StitchingExtensions.graphql"))
+                    .AddSchemaConfiguration(c =>
+                        c.RegisterType<PaginationAmountType>())
+                    .AddExecutionConfiguration(b =>
+                    {
+                        b.AddErrorFilter(error =>
+                            error.AddExtension("STITCH", "SOMETHING"));
+                    }));
+
+            IServiceProvider services =
+                serviceCollection.BuildServiceProvider();
+
+            IQueryExecutor executor = services
+                .GetRequiredService<IQueryExecutor>();
+            IExecutionResult result = null;
+
+            // act
+            using (IServiceScope scope = services.CreateScope())
+            {
+                var request = new QueryRequest(@"
+                {
+                    customer(id: ""Q3VzdG9tZXIteDE="") {
+                        contracts {
+                            id
+                            ... on LifeInsuranceContract {
+                                error
+                            }
+                        }
+                    }
+                }");
+                request.Services = scope.ServiceProvider;
+
+                result = await executor.ExecuteAsync(request);
+            }
 
             // assert
             Snapshot.Match(result);